requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0